requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pydantic==2.6.1
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
pytest==8.0.0
pytest-django==4.7.0
black==24.1.1
flake8==7.0.0
isort==5.13.2 